requirements.txt
asgiref==3.6.0
autopep8==2.0.1
blurhash==1.1.4
CacheControl==0.12.11
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.7
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
